Split oversized modules into focused submodules under 80 logic lines
Refactor large files (parser 466, evaluator 325, type_checker 310,
show_builtin 219, builtin_list 137, value 135 lines) into cohesive
submodules, each holding at most ~80 lines of logic.

- parser: split into atomic/binary/unary/postfix expression builders,
  datetime/duration literal builders, operator mapping, error message
- evaluator: split into binary dispatch, numeric/duration/datetime
  operations, unary, type cast, value conversion
- type_checker: split into binary/cast/function/argument rules
- builtin: split list ops from statistics and extremum
- domain: move calendar helpers to calendar module, Display impl to
  value_display; add Value::type_name and BinaryOperation::symbol to
  remove duplicated error-message helpers
- presentation: separate builtin reference data from printing logic

No behavior change. All 380 tests pass unchanged.
