+++ ccc/domain/src/lib.rs
@@ -1,8 +1,10 @@
 pub mod ast;
+pub mod calendar;
 pub mod error;
 pub mod interface;
 pub mod static_type;
 pub mod value;
+mod value_display;
 
 #[cfg(test)]
 mod ast_test;
