+++ ccc/infrastructure/src/evaluator/mod.rs
@@ -1,5 +1,12 @@
 mod ast_evaluator;
+mod binary_datetime;
+mod binary_duration;
+mod binary_numeric;
+mod binary_operation;
 mod builtin;
+mod type_cast;
+mod unary_operation;
+mod value_conversion;
 
 pub use ast_evaluator::AstEvaluator;
