+++ ccc/infrastructure/src/type_checker/mod.rs
@@ -1,4 +1,8 @@
+mod argument_rule;
 mod ast_type_checker;
+mod binary_rule;
+mod cast_rule;
+mod function_rule;
 
 pub use ast_type_checker::AstTypeChecker;
